[Enhance] refactor PostCardWideReview
