--- src/components/post-list/props.d.ts
@@ -1,5 +1,0 @@
-import ScreenNavigationProps from '@src/modules/types/screen-navigation-props';
-
-type PostListScreenProps = ScreenNavigationProps;
-
-export default PostListScreenProps;
